requirements.txt
pandas
matplotlib
scikit-learn
umap-learn
numpy